Fixed bug that made autozoom trigger when it was not supposed to
More especifically, while using the image strip layout, leaving the window
and coming back to it triggered a size-allocated event in the imageview
widget. If the user had panned the view after setting autozoom it was possible
that the autozoom is turned on but the view is now on a image of different
size. This fixes that error.